make resizable and add support for selected period highlight
