+++ modules/periodo-app/package.json
@@ -38,6 +38,7 @@
     "react-hyperscript": "github:periodo/react-hyperscript",
     "react-redux": "^7.1.0",
     "react-window": "^1.8.3",
+    "react-virtualized-auto-sizer": "^1.0.2",
     "redux": "^3.7.2",
     "respin": "^1.0.1",
     "shallow-equal": "^1.2.0",
